pyproject.toml
[project]
name = "djangodocker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
]
